Enable functionality for custom responses
Holy moly I'm so glad I finally did this. The implementation is nuts
because I have to clone the responses the user chooses in order to be
able to send them multiple times. I ended up writing a bunch of code to
read the response body, clone the data I read, and then write create two
new responses with each dataset. There may or may not be a more
efficient way to do this, but I'll let that be someone else's problem
for the time being.