Begin async/await rewrite
This rewrite is starting with only authorization implemented for now.
Interaction with the Reddit API is possible, but the HTTP requests must
be built themselves, which isn't too hard, and the responses will
not be deserialized other than to plain JSON, which could possibly be
a permanent change since maintaining a copy of the entire Reddit data
model in rust is not something I'm sure I'm up for.